[*] Вынес инициализацию firebase в отдельный модуль, устранил замечания безопасности
