+++ src/App.js
@@ -1,30 +1,21 @@
 import { Route, Switch, useRouteMatch, Redirect } from 'react-router-dom';
 import cn from 'classnames';
-import firebase from 'firebase';
+
+import style from './style.module.css';
 
 import MenuHeader from './components/MenuHeader';
 import Footer from './components/Footer/Footer';
-
 import AboutPage from './routes/About';
 import ContactPage from './routes/Contact';
 import GamePage from './routes/Game';
 import HomePage from './routes/Home';
 import NotFoundPage from './routes/NotFound';
 
-import style from './style.module.css';
+import database from './service/firebase';
 
-const firebaseConfig = {
-  apiKey: "AIzaSyAje9dkqyMdyaOfN0WM-Y7WCh-nNN6iUgU",
-  authDomain: "pokemon-game-f2ef8.firebaseapp.com",
-  databaseURL: "https://pokemon-game-f2ef8-default-rtdb.firebaseio.com",
-  projectId: "pokemon-game-f2ef8",
-  storageBucket: "pokemon-game-f2ef8.appspot.com",
-  messagingSenderId: "636593100078",
-  appId: "1:636593100078:web:36e9eb132b5fb5fd2fd6bf"
-};
-
-firebase.initializeApp(firebaseConfig);
-const database = firebase.database();
+// database.ref('pokemons').once('value', (snapshot) => {
+//   console.log('db data', snapshot.val());
+// })
 
 const App = () => {
   const match = useRouteMatch('/home');
